pyproject.toml
[project]
name = "csnt"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Ftps", email = "ftpsflandre@gmail.com" }]
dependencies = [
    "torch>=2.4.1",
    "pytorch-ignite>=0.5.1",
    "neuron>=8.2.6",
    "omegaconf>=2.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["black>=24.10.0", "isort>=5.13.2", "flake8>=7.1.1"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["csnt"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.black]
extend-exclude = '''
    ^(
        \.venv,
        __pycache__,
    )
'''

[tool.isort]
profile = "black"
skip_gitignore = true
